LPS-56889 Decouple portlet data handler from MBMessage
