+++ modules/apps/message-boards/message-boards-comment/src/com/liferay/message/boards/comment/MBCommentManagerImpl.java
@@ -30,6 +30,7 @@ import com.liferay.portal.security.permission.PermissionChecker;
 import com.liferay.portal.service.ServiceContext;
 import com.liferay.portal.util.PortalUtil;
 import com.liferay.portlet.exportimport.xstream.XStreamAliasRegistryUtil;
+import com.liferay.portlet.messageboards.model.MBCategoryConstants;
 import com.liferay.portlet.messageboards.model.MBDiscussion;
 import com.liferay.portlet.messageboards.model.MBMessage;
 import com.liferay.portlet.messageboards.model.MBMessageDisplay;
@@ -38,6 +39,7 @@ import com.liferay.portlet.messageboards.model.MBTreeWalker;
 import com.liferay.portlet.messageboards.model.impl.MBMessageImpl;
 import com.liferay.portlet.messageboards.service.MBDiscussionLocalService;
 import com.liferay.portlet.messageboards.service.MBMessageLocalService;
+import com.liferay.portlet.messageboards.service.MBThreadLocalService;
 import com.liferay.portlet.messageboards.util.comparator.MessageThreadComparator;
 import com.liferay.portlet.ratings.model.RatingsEntry;
 import com.liferay.portlet.ratings.model.RatingsStats;
@@ -162,6 +164,12 @@ public class MBCommentManagerImpl implements CommentManager {
 		throws PortalException {
 
 		_mbMessageLocalService.deleteDiscussionMessages(className, classPK);
+	}
+
+	@Override
+	public void deleteGroupComments(long groupId) throws PortalException {
+		_mbThreadLocalService.deleteThreads(
+			groupId, MBCategoryConstants.DISCUSSION_CATEGORY_ID);
 	}
 
 	@Override
@@ -315,8 +323,16 @@ public class MBCommentManagerImpl implements CommentManager {
 		XStreamAliasRegistryUtil.unregister(MBMessageImpl.class, "MBMessage");
 	}
 
+	@Reference(unbind = "-")
+	protected void setMBThreadLocalService(
+		MBThreadLocalService mbThreadLocalService) {
+
+		_mbThreadLocalService = mbThreadLocalService;
+	}
+
 	private MBDiscussionLocalService _mbDiscussionLocalService;
 	private MBMessageLocalService _mbMessageLocalService;
+	private MBThreadLocalService _mbThreadLocalService;
 	private RatingsEntryLocalService _ratingsEntryLocalService;
 	private RatingsStatsLocalService _ratingsStatsLocalService;
